backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
